+++ Atomex.Client.Core/Wallet/WalletAddress.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Numerics;
+
 using Atomex.Blockchain;
 using Atomex.Wallet.Abstract;
 
@@ -12,16 +14,16 @@ namespace Atomex.Core
 
         public string Currency { get; set; }
         public string Address { get; set; }
-        public decimal Balance { get; set; }
-        public decimal UnconfirmedIncome { get; set; }
-        public decimal UnconfirmedOutcome { get; set; }
+        public BigInteger Balance { get; set; }
+        public BigInteger UnconfirmedIncome { get; set; }
+        public BigInteger UnconfirmedOutcome { get; set; }
         public KeyIndex KeyIndex { get; set; }
         public bool HasActivity { get; set; }
         public int KeyType { get; set; }
         public TokenBalance TokenBalance { get; set; }
         public DateTime LastSuccessfullUpdate { get; set; }
 
-        public decimal AvailableBalance() => Currencies.IsBitcoinBased(Currency)
+        public BigInteger AvailableBalance() => Currencies.IsBitcoinBased(Currency)
             ? Balance + UnconfirmedIncome
             : Balance;
